move compression-related code to separate file
