+++ lib/commands/client_handshake.js
@@ -105,8 +105,10 @@ ClientHandshake.prototype.handshakeInit = function(helloPacket, connection) {
 ClientHandshake.prototype.handshakeResult = function(okPacket, connection) {
   // error is already checked in base class. Done auth.
   connection.authorized = true;
-  if (connection.config.compress)
-    connection._enableCompression();
+  if (connection.config.compress) {
+    var enableCompression = require('../compressed_protocol.js').enableCompression;
+    enableCompression(connection);
+  }
   // TODO any useful information in ok packet to pass as argument?
   connection.emit('connect', true);
   return null;
