+++ lib/connection.js
@@ -87,29 +87,11 @@ function Connection(opts)
     connection._fatalError = err;
   });
 
-  // big TODO: benchmark if it all worth using 'ondata' and onPacket callbacks directly
-  // compositing streams would be much more easier.
-  // also, look for existing length-prefixed streams to reuse instead of packet_parser
-  //  https://github.com/squaremo/node-spb - currently only fixed 4 byte prefix
-  //  ...?
-
   // see https://gist.github.com/khoomeister/4985691#use-that-instead-of-bind
   this.packetParser = new PacketParser(function(p) { connection.handlePacket(p); });
-  this._inflatedPacketsParser = null;
-
-  // TODO: this code used to be an optimized version of handler
-  // DOES NOT WORK IN NODE 11
-  // TODO: measure if we actually get something here
-  // if yes, re-enable for node 10
-  //if (this.stream instanceof net.Stream) {
-  //  this.stream.ondata = function(data, start, end) {
-  //    connection.packetParser.execute(data, start, end);
-  //  };
-  //} else {
-    this.stream.on('data', function(data) {
-      connection.packetParser.execute(data);
-    });
-  //}
+  this.stream.on('data', function(data) {
+    connection.packetParser.execute(data);
+  });
   this._protocolError = null;
   this.stream.on('end', function() {
     // we need to set this flag everywhere where we want connection to close
@@ -174,25 +156,7 @@ Connection.prototype.write = function(buffer) {
   this.stream.write(buffer);
 };
 
-Connection.prototype._enableCompression = function() {
-  var self = this;
-  this._inflatedPacketsParser = new PacketParser(function(p) {
-    self.handlePacket(p);
-  }, 4);
-  this._inflatedPacketsParser._lastPacket = 0;
-  this.packetParser = new PacketParser(function(packet) {
-    self._handleCompressedPacket(packet);
-  }, 7);
-};
-
-// TODO: replace function in runtime instead of having if() here
-// Needs benchmark.
-
 Connection.prototype.writePacket = function(packet) {
-  var connection = this;
-  var packetSequenceId = this.sequenceId;
-  //console.log('writePacket ====================', packetSequenceId);
-
   packet.writeHeader(this.sequenceId);
   if (this.config.debug) {
     console.log(this._internalId + ' ' + this.connectionId + ' <== ' + this._command._commandName + '#' + this._command.stateName() + '(' + [this.sequenceId, packet._name, packet.length()].join(',') + ')');
@@ -200,39 +164,7 @@ Connection.prototype.writePacket = function(packet) {
   this.sequenceId++;
   if (this.sequenceId == 256)
     this.sequenceId = 0
-
-  if (!this.config.compress || !this.authorized) {
-    this.write(packet.buffer);
-  } else {
-
-    var packetLen = packet.length();
-    var compressHeader = new Buffer(7);
-    var compressed = zlib.deflateSync(packet.buffer);
-    var compressedLength = compressed.length;
-    // NB: to use async zlib.deflate there needs to be code added to prevent race conditions
-    // e.i two writePacket calls can result in second connection.write(compressed) called before first
-    //zlib.deflate(packet.buffer, function(err, compressed) {})...
-
-    if (compressedLength > packetLen + 3) {
-      compressHeader.writeUInt8(compressedLength & 0xff, 0);
-      compressHeader.writeUInt16LE(compressedLength >> 8, 1);
-      compressHeader.writeUInt8(packetSequenceId, 3);
-      compressHeader.writeUInt8(packetLen & 0xff, 4);
-      compressHeader.writeUInt16LE(packetLen >> 8, 5);
-      connection.write(compressHeader);
-      connection.write(compressed);
-    } else {
-      compressedLength = packetLen;
-      packetLen = 0;
-      compressHeader.writeUInt8(compressedLength & 0xff, 0);
-      compressHeader.writeUInt16LE(compressedLength >> 8, 1);
-      compressHeader.writeUInt8(packetSequenceId, 3);
-      compressHeader.writeUInt8(packetLen & 0xff, 4);
-      compressHeader.writeUInt16LE(packetLen >> 8, 5);
-      connection.write(compressHeader);
-      connection.write(packet.buffer);
-    }
-  }
+  this.write(packet.buffer);
 };
 
 Connection.prototype.startTLS = function _startTLS(onSecure) {
@@ -270,58 +202,6 @@ Connection.prototype.startTLS = function _startTLS(onSecure) {
   });
 };
 
-Connection.prototype._checkOOOPackets = function() {
-  var newOOO = [];
-  var connection = this;
-  this._outOfOrderPackets.forEach(function(p) {
-    if (connection._inflatedPacketsParser._lastPacket + 1 === p[0]) {
-      connection._inflatedPacketsParser.execute(p[1]);
-      connection._inflatedPacketsParser._lastPacket = p[0];
-    } else {
-      newOOO.push(p);
-    }
-  });
-  connection._outOfOrderPackets = newOOO;
-};
-
-var zlib = require('zlib');
-var compressedPacketNum = 0;
-Connection.prototype._handleCompressedPacket = function(packet) {
-  compressedPacketNum++;
-  var connection = this;
-  var compressedSequenceId = packet.sequenceId;
-  var deflatedLength = packet.readInt24();
-
-  var body = packet.readBuffer();
-  if (deflatedLength !== 0) {
-    var deflatedBody = zlib.inflateSync(body);
-    connection._inflatedPacketsParser.execute(deflatedBody);
-
-    // because zlib.inflate callback might be fired out of order we'll need to use some
-    // sort of priority queue to keep correct order of executing packets
-    // might not worth efforts ( but potentially beneficial on multi core )
-    /*
-    zlib.inflate(compressedBody, function(err, data) {
-      //console.log('done inflate for ', compressedSequenceId)
-      if (err)
-        return connection.emit('error', err);
-
-      //console.log(connection._inflatedPacketsParser._lastPacket, compressedPacketNum, data.length);
-
-      if (connection._inflatedPacketsParser._lastPacket + 1 == compressedPacketNum) {
-        connection._inflatedPacketsParser._lastPacket = compressedPacketNum;
-        connection._inflatedPacketsParser.execute(data);
-        if (connection._outOfOrderPackets.length > 0)
-          connection._checkOOOPackets();
-      } else {
-        connection._outOfOrderPackets.push([compressedPacketNum, data]);
-      }
-    });
-    */
-  } else {
-    connection._inflatedPacketsParser.execute(body);
-  }
-};
 
 Connection.prototype.pipe = function() {
   var connection = this;
